Added comments and explanations for version 2
Version 2 was added to as a release to version 1:

With more careful attention to the requirements and lectures, this second version of project one is built for mobile first using Flexbox CSS. I decided to use Flexbox as the scruture of the site is simple and therefore it is a more pratical tool for building the site with less code, as apposed to version one which was build using CSS Grid. This version two also includes a person branding and about, contact and home pages. The contact page has a contact form which will need to be hooked up before taking the site live. Also, the navigation has been achieved manually will simple CSS attributes as there are not many pages.